CORE-2671
Updated retailerId column data types to tinyint to reflect changes to
the database.

+++ grails-app/domain/uk/co/wonderlane/wlpos/TillStock.groovy
@@ -22,7 +22,7 @@ class TillStock {
 
         serialNumber column: "serialNumber"
         model column: "model"
-        retailerId column: "retailerId", sqlType: "int"
+        retailerId column: "retailerId", sqlType: "tinyint"
         tillId column: "tillId", sqlType: "int"
         storeId column: "storeId", sqlType: "int"
         dateUpdated column: "dateUpdated"
